+++ package.json
@@ -13,6 +13,7 @@
     "bootstrap": "^5.1.3",
     "dayjs": "^1.11.4",
     "json-server": "^0.17.0",
+    "jwt-decode": "^3.1.2",
     "query-string": "^7.1.1",
     "react": "^18.2.0",
     "react-bootstrap": "^2.4.0",
@@ -21,6 +22,7 @@
     "react-router-dom": "^6.3.0",
     "react-scripts": "5.0.1",
     "react-slick": "^0.29.0",
+    "react-spinner-material": "^1.4.0",
     "sass": "^1.53.0",
     "slick-carousel": "^1.8.1",
     "styled-components": "^5.3.5",
